requirements-section2-2.txt
torch --index-url https://download.pytorch.org/whl/cpu
torchsummary
numpy
pandas
xarray
netCDF4
matplotlib
cartopy
scikit-learn
